+++ src/vue/import.js
@@ -17,6 +17,7 @@ import dog_switch from './out/switch.vue';
 import dog_fold from './out/fold.vue';
 import dog_small from './out/small.vue';
 import dog_float from './out/floatButton.vue';
+import dog_topCredit from './out/topCredit.vue';
 
 export const title = dog_title;
 export const block = dog_block;
@@ -37,3 +38,4 @@ export const switcher = dog_switch;
 export const fold = dog_fold;
 export const small = dog_small;
 export const float = dog_float;
+export const topCredit = dog_topCredit;